+++ requests_auth/authentication.py
@@ -63,7 +63,7 @@ def _get_query_parameter(url: str, param_name: str) -> Optional[str]:
 
 def request_new_grant_with_post(
     url: str, data, grant_name: str, timeout: float, session: requests.Session
-) -> (str, int):
+) -> (str, int, str):
     with session:
         response = session.post(url, data=data, timeout=timeout)
         if not response:
@@ -74,7 +74,7 @@ def request_new_grant_with_post(
     token = content.get(grant_name)
     if not token:
         raise GrantNotProvided(grant_name, content)
-    return token, content.get("expires_in")
+    return token, content.get("expires_in"), content.get('refresh_token')
 
 
 class OAuth2:
@@ -186,17 +186,29 @@ class OAuth2ResourceOwnerPasswordCredentials(requests.auth.AuthBase, SupportMult
             self.data["scope"] = " ".join(scope) if isinstance(scope, list) else scope
         self.data.update(kwargs)
 
+        # As described in https://tools.ietf.org/html/rfc6749#section-6
+        self.refresh_data = {
+            "grant_type": "refresh_token"
+        }
+        if scope:
+            self.refresh_data["scope"] = self.data["scope"]
+        self.refresh_data.update(kwargs)
+
         all_parameters_in_url = _add_parameters(self.token_url, self.data)
         self.state = sha512(all_parameters_in_url.encode("unicode_escape")).hexdigest()
 
     def __call__(self, r):
-        token = OAuth2.token_cache.get_token(self.state, self.request_new_token)
+        token = OAuth2.token_cache.get_token(
+            key=self.state,
+            on_missing_token=self.request_new_token,
+            on_expired_token=self.refresh_token,
+        )
         r.headers[self.header_name] = self.header_value.format(token=token)
         return r
 
     def request_new_token(self):
         # As described in https://tools.ietf.org/html/rfc6749#section-4.3.3
-        token, expires_in = request_new_grant_with_post(
+        token, expires_in, refresh_token = request_new_grant_with_post(
             self.token_url,
             self.data,
             self.token_field_name,
@@ -204,7 +216,19 @@ class OAuth2ResourceOwnerPasswordCredentials(requests.auth.AuthBase, SupportMult
             self.session,
         )
         # Handle both Access and Bearer tokens
-        return (self.state, token, expires_in) if expires_in else (self.state, token)
+        return (self.state, token, expires_in, refresh_token) if expires_in else (self.state, token)
+
+    def refresh_token(self, refresh_token: str):
+        # As described in https://tools.ietf.org/html/rfc6749#section-6
+        self.refresh_data['refresh_token'] = refresh_token
+        token, expires_in, refresh_token = request_new_grant_with_post(
+            self.token_url,
+            self.refresh_data,
+            self.token_field_name,
+            self.timeout,
+            self.session
+        )
+        return self.state, token, expires_in, refresh_token
 
 
 class OAuth2ClientCredentials(requests.auth.AuthBase, SupportMultiAuth):
@@ -262,6 +286,7 @@ class OAuth2ClientCredentials(requests.auth.AuthBase, SupportMultiAuth):
         if scope:
             self.data["scope"] = " ".join(scope) if isinstance(scope, list) else scope
         self.data.update(kwargs)
+        # Refresh tokens are not supported, as described in https://tools.ietf.org/html/rfc6749#section-4.4.3
 
         all_parameters_in_url = _add_parameters(self.token_url, self.data)
         self.state = sha512(all_parameters_in_url.encode("unicode_escape")).hexdigest()
@@ -273,7 +298,7 @@ class OAuth2ClientCredentials(requests.auth.AuthBase, SupportMultiAuth):
 
     def request_new_token(self) -> tuple:
         # As described in https://tools.ietf.org/html/rfc6749#section-4.4.3
-        token, expires_in = request_new_grant_with_post(
+        token, expires_in, _ = request_new_grant_with_post(
             self.token_url,
             self.data,
             self.token_field_name,
@@ -398,8 +423,18 @@ class OAuth2AuthorizationCode(requests.auth.AuthBase, SupportMultiAuth, BrowserA
         }
         self.token_data.update(kwargs)
 
+        # As described in https://tools.ietf.org/html/rfc6749#section-6
+        self.refresh_data = {
+            "grant_type": "refresh_token"
+        }
+        self.refresh_data.update(kwargs)
+
     def __call__(self, r):
-        token = OAuth2.token_cache.get_token(self.state, self.request_new_token)
+        token = OAuth2.token_cache.get_token(
+            key=self.state,
+            on_missing_token=self.request_new_token,
+            on_expired_token=self.refresh_token
+        )
         r.headers[self.header_name] = self.header_value.format(token=token)
         return r
 
@@ -412,7 +447,7 @@ class OAuth2AuthorizationCode(requests.auth.AuthBase, SupportMultiAuth, BrowserA
         # As described in https://tools.ietf.org/html/rfc6749#section-4.1.3
         self.token_data["code"] = code
         # As described in https://tools.ietf.org/html/rfc6749#section-4.1.4
-        token, expires_in = request_new_grant_with_post(
+        token, expires_in, refresh_token = request_new_grant_with_post(
             self.token_url,
             self.token_data,
             self.token_field_name,
@@ -420,7 +455,19 @@ class OAuth2AuthorizationCode(requests.auth.AuthBase, SupportMultiAuth, BrowserA
             self.session,
         )
         # Handle both Access and Bearer tokens
-        return (self.state, token, expires_in) if expires_in else (self.state, token)
+        return (self.state, token, expires_in, refresh_token) if expires_in else (self.state, token)
+
+    def refresh_token(self, refresh_token: str):
+        # As described in https://tools.ietf.org/html/rfc6749#section-6
+        self.refresh_data['refresh_token'] = refresh_token
+        token, expires_in, refresh_token = request_new_grant_with_post(
+            self.token_url,
+            self.refresh_data,
+            self.token_field_name,
+            self.timeout,
+            self.session
+        )
+        return self.state, token, expires_in, refresh_token
 
 
 class OAuth2AuthorizationCodePKCE(
@@ -548,8 +595,18 @@ class OAuth2AuthorizationCodePKCE(
         }
         self.token_data.update(kwargs)
 
+        # As described in https://tools.ietf.org/html/rfc6749#section-6
+        self.refresh_data = {
+            "grant_type": "refresh_token"
+        }
+        self.refresh_data.update(kwargs)
+
     def __call__(self, r):
-        token = OAuth2.token_cache.get_token(self.state, self.request_new_token)
+        token = OAuth2.token_cache.get_token(
+            key=self.state,
+            on_missing_token=self.request_new_token,
+            on_expired_token=self.refresh_token
+        )
         r.headers[self.header_name] = self.header_value.format(token=token)
         return r
 
@@ -562,7 +619,7 @@ class OAuth2AuthorizationCodePKCE(
         # As described in https://tools.ietf.org/html/rfc6749#section-4.1.3
         self.token_data["code"] = code
         # As described in https://tools.ietf.org/html/rfc6749#section-4.1.4
-        token, expires_in = request_new_grant_with_post(
+        token, expires_in, refresh_token = request_new_grant_with_post(
             self.token_url,
             self.token_data,
             self.token_field_name,
@@ -570,7 +627,20 @@ class OAuth2AuthorizationCodePKCE(
             self.session,
         )
         # Handle both Access and Bearer tokens
-        return (self.state, token, expires_in) if expires_in else (self.state, token)
+        return (self.state, token, expires_in, refresh_token) if expires_in else (self.state, token)
+
+    def refresh_token(self, refresh_token: str):
+        # As described in https://tools.ietf.org/html/rfc6749#section-6
+        self.refresh_data['refresh_token'] = refresh_token
+        token, expires_in, refresh_token = request_new_grant_with_post(
+            self.token_url,
+            self.refresh_data,
+            self.token_field_name,
+            self.timeout,
+            self.session
+        )
+        return self.state, token, expires_in, refresh_token
+
 
     @staticmethod
     def generate_code_verifier() -> bytes: