+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for refresh tokens in the Resource Owner Password Credentials flow and Authorization Code (with and without PKCE) flows.
 
 ## [5.2.0] - 2020-10-14
 ### Added